ob: Support for detecting errors in sessions
* lisp/ob-eval.el (org-babel-session-error-value): New constant. A string that
will be used to signal that an error has occurred in a session.
(org-babel-session-error-langs): New variable. List of languages for
which errors in sessions should be detected.
(org-babel-session-error-regexp): New function: Constructs the regexp which
tests whether a result indicates error, and parses this result into an
error code and an error message.